Refactor api and hook file structure to make more sense, renamed Compare Candidates tab to Bills
