+++ politiprofile/hooks/user/useGetUserDetails.js
@@ -12,7 +12,7 @@ const useGetUserDetails = () => {
       if (user) {
         console.log("Fetching details for user:", user.email); // Log user email for debugging
         try {
-          const response = await fetch("/api/db/getUserDetails", {
+          const response = await fetch("/api/user/getUserDetails", {
             method: "POST",
             headers: {
               "Content-Type": "application/json",
