+++ politiprofile/pages/api/user/getUserDetails.js
